requirements.txt
pandas
schedule
pvlib==0.8.0
datetime
pillow
smbus
pygame